enclave/app/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
pydantic==2.11.7
cryptography==45.0.4
pycryptodome==3.23.0
